Refactor slightly to make it more unreadable
